CRUD operations tests now bypass authentification
