+++ solr-admin-app/tests/conftest.py
@@ -12,7 +12,7 @@ def port():
     return 8080
 
 
-@pytest.fixture(scope="session")
+@pytest.fixture(scope="function")
 def server_with_real_keycloak(port):
     app = os.path.join(os.path.dirname(__file__), '..', 'app.py')
     server = ServerDriver(name='MyServer', port=port)
@@ -20,9 +20,9 @@ def server_with_real_keycloak(port):
     return server
 
 
-@pytest.fixture(scope="session")
+@pytest.fixture(scope="function")
 def server_with_fake_keycloak(port):
-    app = os.path.join(os.path.dirname(__file__), '..', 'app_test.py')
+    app = os.path.join(os.path.dirname(__file__), '..', 'tests', 'external', 'support', 'app_test.py')
     server = ServerDriver(name='MyServer', port=port)
     server.start(cmd=['python', app])
     return server
@@ -41,19 +41,19 @@ def get_browser():
 
 
 @pytest.fixture(scope="function")
-def browser_against_fake_keycloak(server_with_fake_keycloak):
-    browser = get_browser()
-    yield browser
-    #browser.quit()
-    server_with_fake_keycloak.shutdown()
-
-
-@pytest.fixture(scope="function")
 def browser_against_real_keycloak(server_with_real_keycloak):
     browser = get_browser()
     yield browser
-    #browser.quit()
+    browser.quit()
     server_with_real_keycloak.shutdown()
+
+
+@pytest.fixture(scope="function")
+def browser(server_with_fake_keycloak):
+    browser = get_browser()
+    yield browser
+    browser.quit()
+    server_with_fake_keycloak.shutdown()
 
 
 @pytest.fixture(scope="session")
@@ -69,7 +69,10 @@ def clean_db():
     from solr_admin.models.restricted_condition import RestrictedCondition
     from solr_admin.models.restricted_word import RestrictedWord
     from solr_admin.models.restricted_word_condition import RestrictedWordCondition
+    from tests.external.support.fake_oidc import FakeOidc
+    from solr_admin.keycloak import Keycloak
 
+    Keycloak._oidc = FakeOidc()
     app, admin = create_application(run_mode='testing')
 
     db = SQLAlchemy(app)
